Update touch listeners for charts (detect horizontal scroll)

+++ app/src/main/java/com/kitzapp/telegram_stats/presentation/ui/components/ChartView/impl/ViewChartMiniature.java
@@ -74,7 +74,7 @@ public class ViewChartMiniature extends ViewChartBase {
 
     @Override
     public void update(Observable o, Object arg) {
-        if ((int) arg == ObserverManager.KEY_OBSERVER_THEME_UPDATED) {
+        if ((byte) arg == ObserverManager.KEY_OBSERVER_THEME_UPDATED) {
             int newFullChartBackColor = getFullChartBackColor();
 
             // FULL CHART CHANGE BACK COLOR
