+++ app/src/main/java/com/kitzapp/telegram_stats/presentation/ui/components/simple/TDelimiterLine.java
@@ -64,7 +64,7 @@ public class TDelimiterLine extends View implements TViewObserver {
 
     @Override
     public void update(Observable o, Object arg) {
-        if ((int) arg == ObserverManager.KEY_OBSERVER_THEME_UPDATED) {
+        if ((byte) arg == ObserverManager.KEY_OBSERVER_THEME_UPDATED) {
             int newBackColor = getCurrentBackColor();
 
             if (_oldBackColor != newBackColor) {
